Add support for Livestatus Backend
I added the neccessary functions to support the livestatus backend.
The changes in marker.php have been kept to a minimum.
I would prefer to completely seperate the information handling from
marker.php, so we just provide marker.php with an array of prefiltered
hosts to add to the map.